LPD-88102 Rename preview services to match their exported function
